feat: move debounced logic to a helper
